+++ commands/clear.js
@@ -10,8 +10,12 @@ module.exports = {
         if(args[0] > 100) return message.reply("Du kannst nur maximal 100 Nachrichten löschen!");
         if(args[0] < 1) return message.reply("Du musst mindestens eine Nachricht löschen!");
 
-        await message.channel.messages.fetch({limit: args[0]}).then(messages =>{
-            message.channel.bulkDelete(messages);
-        });
+        if(message.member.roles.cache.some(r=>["Admin", "Moderator", "IT-spezialist"].includes(r.name))) {
+            await message.channel.messages.fetch({limit: args[0]}).then(messages =>{
+                message.channel.bulkDelete(messages);
+            });
+        } else {
+            message.channel.send("Du hast nicht genügend Rechte für diesen Command!");
+        }
     }
 }
